requirements.txt
#Adafruit_DHT
paho-mqtt
#sqlite3
matplotlib
#random